fix: Allow for more than 255 container restarts in forge-dpu-agent
dpu-agent deserializes various container status field - including the `attempt` field. When the field reached a value of > 255, dpu-agent failed to deserialize the field. This change prevent this.

Fixes https://nvbugspro.nvidia.com/bug/4868990

See merge request nvmetal/carbide!2960

+++ bluefield/agent/src/containerd/container.rs
@@ -11,7 +11,7 @@ use crate::containerd::{BashCommand, Command};
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct ContainerMetadata {
     pub name: String,
-    pub attempt: u8,
+    pub attempt: u64,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
